Update names of festivals and Edinburgh Festivals Listings API
Updated the text to reflect the correct names for the festival
datasets that are being used. Also removed reference to applying
for access to the Edinburgh Fringe dataset to avoid confusion -
the Book Festival data used in the course just requires
self-service registration for an API key and does not need
additional approval.